+++ src/pytestarch/eval_structure/types.py
@@ -8,7 +8,7 @@ class Import(ABC):
     def __init__(self, importer: str) -> None:
         self._importer = importer
 
-        self._importer_module_hierarchy = self._get_parent_modules(self._importer)
+        self._importer_module_hierarchy = get_parent_modules(self._importer)
 
     def importer(self) -> str:
         """Returns name of the module that imports something.
@@ -47,30 +47,31 @@ class Import(ABC):
     def __str__(self) -> str:
         return f"{self.importer()} imports {self.importee()}"
 
-    def _get_parent_modules(self, module: str) -> List[str]:
-        """Calculates all parent modules of a given module.
-
-        Example: source root is a
-        module: a.b.c
-        returned: [a, a.b]
-
-        Args:
-            module: module to calculate parent modules for
-
-        Returns:
-            List of all parent modules, containing their full names up to the source code root.
-        """
-        parent_modules = []
-
-        parent: List[str] = []
-
-        for char in module:
-            if char == ".":
-                parent_modules.append("".join(parent))
-
-            parent.append(char)
-
-        return parent_modules
-
     def _append_prefix(self, value: str, prefix: str) -> str:
         return f"{prefix}.{value}"
+
+
+def get_parent_modules(module: str) -> List[str]:
+    """Calculates all parent modules of a given module.
+
+    Example: source root is a
+    module: a.b.c
+    returned: [a, a.b]
+
+    Args:
+        module: module to calculate parent modules for
+
+    Returns:
+        List of all parent modules, containing their full names up to the source code root.
+    """
+    parent_modules = []
+
+    parent: List[str] = []
+
+    for char in module:
+        if char == ".":
+            parent_modules.append("".join(parent))
+
+        parent.append(char)
+
+    return parent_modules